重构 qBittorrent 相关代码，允许 qbEE 继承 qb 代码；清理 v4.x private torrent 相关代码
