+++ src/main/java/com/ghostchu/peerbanhelper/PeerBanHelperServer.java
@@ -8,8 +8,8 @@ import com.ghostchu.peerbanhelper.downloader.DownloaderLastStatus;
 import com.ghostchu.peerbanhelper.downloader.DownloaderLoginResult;
 import com.ghostchu.peerbanhelper.downloader.impl.biglybt.BiglyBT;
 import com.ghostchu.peerbanhelper.downloader.impl.deluge.Deluge;
-import com.ghostchu.peerbanhelper.downloader.impl.qbittorrent.QBittorrent;
-import com.ghostchu.peerbanhelper.downloader.impl.qbittorrent.QBittorrentEE;
+import com.ghostchu.peerbanhelper.downloader.impl.qbittorrent.impl.QBittorrent;
+import com.ghostchu.peerbanhelper.downloader.impl.qbittorrent.impl.enhanced.QBittorrentEE;
 import com.ghostchu.peerbanhelper.downloader.impl.transmission.Transmission;
 import com.ghostchu.peerbanhelper.event.LivePeersUpdatedEvent;
 import com.ghostchu.peerbanhelper.event.PBHServerStartedEvent;
